requirements.txt
Django>=6.0,<6.11
gunicorn>=23.0,<24.0
PyMySQL>=1.1,<2.0

websockets
aiortc
djangorestframework
boto3
python-dotenv

# STT (OpenAI Whisper API)
openai>=1.0

# LLM (Google Gemini)
google-genai

# TTS (Google Cloud)
google-cloud-texttospeech>=2.16

# Audio processing
numpy

# PDF text extraction
pymupdf
